Renamed cronosd folder to genesisd
