+++ .github/workflows/lint.yml
@@ -21,7 +21,7 @@ jobs:
     steps:
       - uses: actions/setup-go@v3
         with:
-          go-version: '^1.20.0'
+          go-version: "^1.20.0"
       - uses: actions/checkout@v3
       - uses: cachix/install-nix-action@v22
         with:
@@ -48,7 +48,7 @@ jobs:
 
           BUILD_TAGS=rocksdb,grocksdb_clean_link
           echo "go build"
-          go build -tags $BUILD_TAGS ./cmd/cronosd
+          go build -tags $BUILD_TAGS ./cmd/genesisd
           echo "golangci-lint ."
           golangci-lint run --out-format=github-actions --path-prefix=./ --timeout 10m --build-tags $BUILD_TAGS
         # Check only if there are differences in the source code
