+++ cmd/genesisd/cmd/genaccounts.go
